requirements.txt
PyPDF2
pdf2image
#fitz
pdfplumber
frontend
pymupdf
pillow
ollama
google-generativeai